requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cryptography==44.0.0
cymem==2.0.11
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
nltk==3.9.1
numpy==2.2.2
openai==1.59.7
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
preshed==3.0.9
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
setuptools==75.8.0
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
textblob==0.19.0
thinc==8.3.4
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
whitenoise==6.8.2
wrapt==1.17.2
